[Behat][Checkout] Raw moving addressing steps
